core: Initial support for native list variables
This commit adds support for native list variables and outputs, building
up on the previous change to state. Interpolation functions currently
are not working, so there are many real-world situations which will
fail, however the tests are all passing right now since they still use
StringList and split/join where necessary.

List variables are defined like this:

```
variable "test" {
    # This can also be inferred
    type = "list"
    default = ["Hello", "World"]
}

output "test_out" {
    value = "${var.a_list}"
}
```

This results in the following state:

```
...
            "outputs": {
                "test_out": [
                    "hello",
                    "world"
                ]
            },
...
```

And the result of `terraform output` is as follows:

```
$ terraform output
test_out = [hello world]
```

The formatting of outputs on the command line requires some thought as
to how to make them friendly for additional processing with
sed/awk/whatever - this will be addressed in a later commit.